% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream-base-env.R
\name{create_stream}
\alias{create_stream}
\title{An environment object encapsulating a stream of tokens and functions for manipulating/interrogating these token.}
\usage{
create_stream(named_values)
}
\arguments{
\item{named_values}{named vector containing the tokens. Usually the output
from \code{lex()}}
}
\description{
This is very similar to the R6 Class \code{TokenStream}, but it has no
dependencies
}
