% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex.R
\docType{data}
\name{regex}
\alias{regex}
\title{Regexes to match common elements}
\format{
An object of class \code{list} of length 3.
}
\usage{
regex
}
\description{
Regexes to match common elements
}
\keyword{datasets}
